Created MariogatchiFactory and EnvironmentFactory to handle enums in the request models. The forest subclass is now able to generate 1-5 new Mariogatchis when a new instance is created. Further information regarding the specifications of the Mariogatchis created can be found in the Forest class. User now has an environment attribute. Imported some things for the add Mariogatchi use case. Added more meaningful labels to the entites/interfaces/factories etc.
Further implementation:
- add an iterator for the find/add Mariogatchi use case
bit.ly/3BiX9ea - github repo
https://www.javatpoint.com/iterator-pattern - external site
- create test cases (good luck) regarding change environment use case, Mariogatchi creation (?), Environment creation (?), and find/add Mariogatchi use case after implementation is complete

Notes:
- Make sure Mariogatchi has either a default constructor or a constructor for just rarity.
- Consider making Park and Forest singleton classes
